Created a behavior in order to move reference to UI out of DockLayoutViewModel
